requirements.txt
aiohttp==3.9.3
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.2.0
Brotli==1.1.0
brotli-asgi==1.4.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-cache2==0.2.1
fastapi-jwt-auth==0.5.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
nanoid==2.0.0
orjson==3.9.13
packaging==23.2
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
